feat:流水线支持展示运行进度 #7932

+++ src/backend/ci/core/worker/worker-common/src/main/kotlin/com/tencent/devops/worker/common/utils/CommandLineUtils.kt
@@ -183,7 +183,7 @@ object CommandLineUtils {
             if (taskId != null) {
                 Heartbeat.recordTaskProgressRate(
                     taskId = taskId,
-                    progressRate = progressRate
+                    progressRate = progressRate.toDouble()
                 )
             }
             logger.info("report progress rate:$tmpLine|$taskId|$progressRate")
